+++ db/migrate/20230902220518_add_site_and_kpi_to_account.rb
@@ -1,0 +1,6 @@
+class AddSiteAndKpiToAccount < ActiveRecord::Migration[7.0]
+  def change
+    add_column :accounts, :site, :integer
+    add_column :accounts, :enable_kpi, :boolean, default: true
+  end
+end