+++ Changelog.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 
+## [0.3.0] : 2023-04-09
+
+### Added
+
+- Added site, enable kpi, status attributes on account
+- Redesign accounts index from cards to table
+- Implement pagy and filter by site
+
 ## [0.2.0] : 2023-03-09
 
 ### Added
@@ -36,6 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 [unreleased]: https://github.com/danrayfr/unified/branch/development#diff
+[0.3.0]: https://github.com/danrayfr/unified/pull/6
+[0.2.0]: https://github.com/danrayfr/unified/pull/5
 [0.1.2]: https://github.com/danrayfr/unified/pull/3
 [0.1.1]: https://github.com/danrayfr/unified/pull/1
 [0.1.0]: https://github.com/danrayfr/unified